+++ NEWS.md
@@ -1,3 +1,4 @@
+* Created `eq_clean_data()` function and corresponding tests.
 * Created `eq_location_clean()` function and corresponding tests.
 * Added support for tibbles.
 * NOAA raw-data imported and documented into the package data like a